remove old WSO2APIPublisher and rename WSO2PublisherNew -> WSO2APIPublisher
