+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/APIProviderImpl.java
@@ -103,7 +103,6 @@ import org.wso2.carbon.apimgt.impl.notification.NotificationExecutor;
 import org.wso2.carbon.apimgt.impl.notification.NotifierConstants;
 import org.wso2.carbon.apimgt.impl.notification.exception.NotificationException;
 import org.wso2.carbon.apimgt.impl.publishers.WSO2APIPublisher;
-import org.wso2.carbon.apimgt.impl.publishers.WSO2PublisherNew;
 import org.wso2.carbon.apimgt.impl.template.APITemplateBuilder;
 import org.wso2.carbon.apimgt.impl.template.APITemplateBuilderImpl;
 import org.wso2.carbon.apimgt.impl.template.APITemplateException;
@@ -3669,7 +3668,7 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
                 String version = ApiMgtDAO.getInstance().getLastPublishedAPIVersionFromAPIStore(api.getId(),
                         store.getName());
 
-                if (apiOlderVersionExist && version != null && !(publisher instanceof WSO2PublisherNew)) {
+                if (apiOlderVersionExist && version != null && !(publisher instanceof WSO2APIPublisher)) {
                     published = publisher.createVersionedAPIToStore(api, store, version);
                     publisher.updateToStore(api, store);
                 } else {
