+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/test/java/org/wso2/carbon/apimgt/impl/publishers/WSO2APIPublisherWrapper.java
@@ -21,18 +21,13 @@ package org.wso2.carbon.apimgt.impl.publishers;
 import org.apache.http.client.HttpClient;
 import org.wso2.carbon.apimgt.api.APIProvider;
 
-public class WSO2APIPublisherWrapper extends WSO2PublisherNew {
+public class WSO2APIPublisherWrapper extends WSO2APIPublisher {
     private HttpClient client;
     private String username;
     private APIProvider apiProvider;
 
     public WSO2APIPublisherWrapper(HttpClient client, String username, APIProvider apiProvider) {
         this.client = client;
-        this.username = username;
-        this.apiProvider = apiProvider;
-    }
-
-    public WSO2APIPublisherWrapper(String username, APIProvider apiProvider) {
         this.username = username;
         this.apiProvider = apiProvider;
     }
